compositor|refactor: Split the config schemas into three modules
The two declaration blocks a tier body composes now have modules of their own, `selection-schemas.ts` and `source-declaration-schemas.ts`, leaving `config-schemas.ts` holding the tier and the config envelope. Every schema stays published under the name it had.
