+++ packages/compositor/src/schemas/config-schemas.ts
@@ -11,102 +11,9 @@
 
 import { z } from 'zod';
 
-import { compareStrings } from '../portable/compareStrings.ts';
-import { SourceOriginSchema } from './descriptor-schemas.ts';
 import { IdSchema } from './scalar-schemas.ts';
-
-/**
- * One `use` or `drop` entry: an artifact named by slug, or everything a source carries.
- *
- * A bare string is the artifact form, so the common case stays terse. The object forms are told apart by which key is
- * present rather than by a sentinel prefix, which would reserve a character out of the slug namespace. An entry naming
- * both is rejected rather than settled by precedence: the key each form does not own is typed `never`, so a mistake
- * fails the parse instead of being silently dropped by the strip that would otherwise swallow it.
- */
-export const SelectorSchema = z.union([
-  z
-    .string()
-    .min(1)
-    .transform((artifact) => ({ artifact })),
-  z.object({ artifact: z.string().min(1), source: z.never().optional() }).transform(({ artifact }) => ({ artifact })),
-  z.object({ source: z.string().min(1), artifact: z.never().optional() }).transform(({ source }) => ({ source })),
-]);
-
-/**
- * One declared content source, in either the authored or the normalized spelling.
- *
- * `path` and `package` are the two authored spellings of the plan schema's source origins, and `origin` is what they
- * normalize to. A source declares exactly one of the three. The location stays as authored: resolving it against the
- * declaring tier is `resolveSources`' job, so a plan reports the path a consumer wrote rather than where it landed.
- */
-export const DeclaredSourceSchema = z.union([
-  z
-    .object({
-      name: z.string().min(1),
-      path: z.string().min(1),
-      package: z.never().optional(),
-      origin: z.never().optional(),
-    })
-    .transform(({ name, path }) => ({ name, origin: { kind: 'directory' as const, location: path } })),
-  z
-    .object({
-      name: z.string().min(1),
-      package: z.string().min(1),
-      path: z.never().optional(),
-      origin: z.never().optional(),
-    })
-    .transform(({ name, package: location }) => ({ name, origin: { kind: 'package' as const, location } })),
-  z
-    .object({
-      name: z.string().min(1),
-      origin: SourceOriginSchema,
-      path: z.never().optional(),
-      package: z.never().optional(),
-    })
-    .transform(({ name, origin }) => ({ name, origin })),
-]);
-
-/**
- * One tier's source declarations: the sources it adds, and the inherited names it drops.
- *
- * `drop` names sources rather than carrying entries, because a name is the whole of what identifies one. Dropping is
- * what the port could do to a package and not to a source; unifying the two lists keeps the stronger semantics.
- */
-export const SourceDeclarationSchema = z.strictObject({
-  use: z.array(DeclaredSourceSchema).default([]),
-  drop: z.array(z.string().min(1)).default([]),
-});
-
-/** One kind's selections within a tier. */
-export const KindSelectionSchema = z.strictObject({
-  kindId: IdSchema,
-  use: z.array(SelectorSchema).default([]),
-  drop: z.array(SelectorSchema).default([]),
-});
-
-/**
- * A tier's selections across every kind it speaks about, as a `kindId`-keyed mapping or as the normalized array.
- *
- * The array form is tried first, so an array input never reaches the record branch. Entries sort by `kindId`, which is
- * what keeps a config authored in two different orders digesting to one value.
- */
-export const SelectSchema = z
-  .union([
-    z.array(KindSelectionSchema),
-    z
-      .record(
-        IdSchema,
-        z.preprocess(
-          (value) => value ?? undefined,
-          KindSelectionSchema.omit({ kindId: true }).default({ use: [], drop: [] }),
-        ),
-      )
-      .transform((blocks) => Object.entries(blocks).map(([kindId, block]) => ({ kindId, ...block }))),
-  ])
-  .refine((blocks) => new Set(blocks.map(({ kindId }) => kindId)).size === blocks.length, {
-    error: 'names a kind more than once',
-  })
-  .transform((blocks) => blocks.toSorted((left, right) => compareStrings(left.kindId, right.kindId)));
+import { SelectSchema } from './selection-schemas.ts';
+import { SourceDeclarationSchema } from './source-declaration-schemas.ts';
 
 /**
  * What one tier file carries.
@@ -151,8 +58,4 @@ export const CompositorConfigSchema = z.strictObject({
 
 export type CompositorConfig = z.infer<typeof CompositorConfigSchema>;
 export type ConfigTier = z.infer<typeof ConfigTierSchema>;
-export type DeclaredSource = z.infer<typeof DeclaredSourceSchema>;
-export type KindSelection = z.infer<typeof KindSelectionSchema>;
-export type Selector = z.infer<typeof SelectorSchema>;
-export type SourceDeclaration = z.infer<typeof SourceDeclarationSchema>;
 export type TierBody = z.infer<typeof TierBodySchema>;
